remove prisma
