+++ .env.example
@@ -1,11 +1,3 @@
-# Postgres
-# Nest run locally
-DATABASE_URL="postgresql://postgres:<REPLACE_ME>@localhost:5432/postgres?schema=public&sslmode=prefer"
-# Nest run in docker, change host to database container name
-# DATABASE_URL="postgresql://postgres:<REPLACE_ME>@cinepik-catalog-db:5432/postgres?schema=public&sslmode=prefer"
-DB_PORT=5432
-POSTGRES_PASSWORD="<REPLACE_ME>"
-
 # Nest
 NODE_PORT=3001
 # Set NODE ENV to development or production
