+++ Dockerfile
@@ -6,9 +6,6 @@ WORKDIR /app
 # Copy package.json and package-lock.json
 COPY package*.json tsconfig*.json ./
 
-# Copy prisma schema
-COPY prisma ./prisma/
-
 # Install dependencies
 RUN npm ci
 
@@ -28,15 +25,12 @@ COPY --from=build /app/node_modules ./node_modules
 COPY --from=build /app/package*.json ./
 COPY --from=build /app/tsconfig*.json ./
 COPY --from=build /app/dist ./dist
-COPY --from=build /app/prisma ./prisma
 
 # Expose the port the app runs on
 EXPOSE ${NODE_PORT}
 
-# Run prisma generate as an additional safety measure
-RUN npx prisma generate
 
 USER node
 
 # Command to run the application
-CMD npm run start:migrate:prod
+CMD npm run start:prod
