+++ README.md
@@ -39,27 +39,6 @@ npm run start:dev
 npm run start:prod
 ```
 
-## Prisma
-
-If any changes are made to the schema.prisma file, run the following command to update the database.
-It creates a migration, updates the database (also seeds?), and updates the Prisma client.
-
-```bash
-npm run prisma:migrate:dev
-```
-
-If you only want to seed the database, run the following command.
-
-```bash
-npm run prisma:seed
-```
-
-To update the Prisma client run the following command.
-
-```bash
-npm run prisma:client
-```
-
 ## Test
 
 ```bash
