+++ docker-compose.yml
@@ -16,35 +16,10 @@ services:
     networks:
       - cinepik-network
 
-  seed:
-    container_name: cinepik-catalog-seed
-    build: .
-    command: npm run prisma:seed
-    depends_on:
-      - db
-    volumes:
-      - .:/app
-      - node_modules:/app/node_modules
-    env_file:
-      - .env
-    networks:
-      - cinepik-network
-
-  db:
-    container_name: cinepik-catalog-db
-    image: postgres:15.5-alpine
-    ports:
-      - '${DB_PORT}:${DB_PORT}'
-    volumes:
-      - postgres_data:/var/lib/postgresql/data
-    env_file:
-      - .env
-    networks:
-      - cinepik-network
-
 volumes:
   postgres_data:
   node_modules:
+
 
 networks:
   cinepik-network:
