--- k8s/seed-job.yml
@@ -1,19 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: cinepik-catalog-seeding-job
-spec:
-  template:
-    spec:
-      containers:
-        - name: cinepik-catalog-seeder
-          image: sebam/cinepik-catalog:latest
-          command: ['npm', 'run', 'prisma:seed']
-          env:
-            - name: DATABASE_URL
-              valueFrom:
-                secretKeyRef:
-                  name: database-credentials
-                  key: DATABASE_URL
-      restartPolicy: Never
-  backoffLimit: 2
