--- prisma/schema.prisma
@@ -1,41 +1,0 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-generator client {
-  provider      = "prisma-client-js"
-  seed          = "prisma/seed.ts"
-  binaryTargets = ["native", "linux-musl-arm64-openssl-3.0.x"]
-}
-
-datasource db {
-  provider = "postgresql"
-  url      = env("DATABASE_URL")
-}
-
-model Movie {
-  id               Int      @id @default(autoincrement())
-  title            String
-  description      String?
-  releaseDate      DateTime
-  director         String
-  genres           String[]
-  rating           Float
-  runtimeMinutes   Int
-  createdAt        DateTime @default(now())
-  updatedAt        DateTime @updatedAt
-}
-
-model Series {
-  id               Int      @id @default(autoincrement())
-  title            String
-  description      String?
-  releaseDate      DateTime
-  creator          String
-  genres           String[]
-  seasons          Int
-  episodes         Int
-  rating           Float
-  runtimeMinutes   Int
-  createdAt        DateTime @default(now())
-  updatedAt        DateTime @updatedAt
-}
