+++ src/app.module.ts
@@ -3,7 +3,7 @@ import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
 import { MoviesModule } from './movies/movies.module';
-import { SeriesModule } from './series/series.module';
+import { ShowsModule } from './shows/shows.module';
 import { KeycloakModule } from './keycloak/keycloak.module';
 import { HealthModule } from './health/health.module';
 import { CommonModule } from './common/common.module';
@@ -16,7 +16,7 @@ import { CommonModule } from './common/common.module';
     }),
     KeycloakModule,
     MoviesModule,
-    SeriesModule,
+    ShowsModule,
     HealthModule,
     CommonModule,
   ],
