+++ src/health/health.controller.ts
@@ -5,9 +5,7 @@ import {
   HealthCheckService,
   HttpHealthIndicator,
   MemoryHealthIndicator,
-  PrismaHealthIndicator,
 } from '@nestjs/terminus';
-import { PrismaService } from 'src/prisma/prisma.service';
 import { ManualHealthIndicator } from './manual.health';
 import { ApiTags } from '@nestjs/swagger';
 import { Unprotected } from 'nest-keycloak-connect';
@@ -18,8 +16,6 @@ export class HealthController {
   constructor(
     private health: HealthCheckService,
     private http: HttpHealthIndicator,
-    private prisma: PrismaHealthIndicator,
-    private prismaService: PrismaService,
     private readonly disk: DiskHealthIndicator,
     private memory: MemoryHealthIndicator,
     private manualHealthIndicator: ManualHealthIndicator,
@@ -41,7 +37,6 @@ export class HealthController {
     // TODO: Add HTTP health check to ping external API
     return this.health.check([
       () => this.http.pingCheck('google', 'https://google.com'),
-      () => this.prisma.pingCheck('prisma', this.prismaService),
       () =>
         this.disk.checkStorage('storage', { path: '/', thresholdPercent: 0.9 }), // if more than 90% of disk space is used
       () => this.memory.checkHeap('memory_heap', 256 * 1024 * 1024), // if more than 256MiB
