+++ src/health/health.module.ts
@@ -3,11 +3,10 @@ import { HttpModule } from '@nestjs/axios';
 import { TerminusModule } from '@nestjs/terminus';
 import { HealthController } from './health.controller';
 import { ManualHealthIndicator } from './manual.health';
-import { PrismaService } from 'src/prisma/prisma.service';
 
 @Module({
   imports: [TerminusModule, HttpModule],
   controllers: [HealthController],
-  providers: [ManualHealthIndicator, PrismaService],
+  providers: [ManualHealthIndicator],
 })
 export class HealthModule {}
