--- src/series/series.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { SeriesController } from './series.controller';
-import { SeriesService } from './series.service';
-import { PrismaService } from 'src/prisma/prisma.service';
-import { KeycloakModule } from 'src/keycloak/keycloak.module';
-
-@Module({
-  imports: [KeycloakModule],
-  controllers: [SeriesController],
-  providers: [SeriesService, PrismaService],
-})
-export class SeriesModule {}
