+++ src/shows/shows.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { SeriesController } from './shows.controller';
+import { ShowsService } from './shows.service';
+import { KeycloakModule } from 'src/keycloak/keycloak.module';
+
+@Module({
+  imports: [KeycloakModule],
+  controllers: [SeriesController],
+  providers: [ShowsService],
+})
+export class ShowsModule {}
