+++ tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["node_modules", "test", "dist", "**/*spec.ts", "prisma"]
+  "exclude": ["node_modules", "test", "dist", "**/*spec.ts"]
 }
